requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
sqlalchemy
jupyter
python-dotenv
groq
streamlit
requests
